Refactor common code out of test_config_schema
