+++ test/config/test_config_schema.py
@@ -1,21 +1,13 @@
-import os
 import pytest
 import yaml
-import glob
 from jsonschema import validate
+from utils import configs
 
 
 @pytest.fixture
 def schema():
     with open("src/config/schema.yml") as f:
         return yaml.safe_load(f)
-
-
-def configs():
-    for config_path in glob.glob("examples/config/*.yml"):
-        if os.path.isfile(config_path):
-            with open(config_path) as f:
-                yield yaml.safe_load(f), config_path
 
 
 @pytest.mark.parametrize("config", configs(), ids=lambda c: c[1])
